// src/types/udbhavForm.ts
export type udbhavForm = {
    category: string;
    subCategory: string;
    participants: string[];
    school: string;
    contact: string;
    email: string;
    class: string;
    age: string;
    choreographer: string;
    paymentScreenshot: File | null;
    confirmDetails: boolean;
    agreeRules: boolean;
    agreeCancellation: boolean;
};